+++ R/check_predictions.R
@@ -37,6 +37,9 @@
 #' `"discrete_dots"`, `"discrete_interval"` or `"discrete_both"` (the `discrete_*`
 #' options are appropriate for models with discrete - binary, integer or ordinal
 #' etc. - outcomes).
+#' @param x_limits An integer vector of length two specifying the x-axis limits
+#' for the plot. Use this to zoom in on a specific region of interest,
+#' especially if the response variable has a large range.
 #' @param verbose Toggle warnings.
 #' @param ... Additional arguments passed on to downstream functions. For
 #' frequentist models, these are forwarded to `simulate()`; for Bayesian models
@@ -111,6 +114,7 @@ check_predictions.default <- function(
   re_formula = NULL,
   bandwidth = "nrd",
   type = "density",
+  x_limits = NULL,
   verbose = TRUE,
   object = NULL,
   ...
@@ -162,6 +166,7 @@ check_predictions.default <- function(
     type = type,
     verbose = verbose,
     model_info = minfo,
+    x_limits = x_limits,
     ...
   )
 }
@@ -329,6 +334,7 @@ pp_check.lm <- function(
   type = "density",
   verbose = TRUE,
   model_info = NULL,
+  x_limits = NULL,
   ...
 ) {
   # we need the formula and the response values to check for matrix responses
@@ -347,12 +353,28 @@ pp_check.lm <- function(
       type,
       verbose,
       model_info,
+      x_limits,
       ...
     ))
   }
 
   # else, proceed as usual
   out <- .safe(stats::simulate(object, nsim = iterations, re.form = re_formula, ...))
+
+  # if it fails, try insight methods
+  if (is.null(out)) {
+    out <- .safe({
+      sims <- insight::get_simulated(
+        object,
+        iterations = iterations,
+        re.form = re_formula,
+        ...
+      )
+      # get_simulated returns "iter_" columms, so we rename here
+      colnames(sims) <- gsub("^iter_(\\d+)", "sim_\\1", colnames(sims))
+      sims
+    })
+  }
 
   # validation check, for mixed models, where re.form = NULL (default) might fail
   out <- .check_re_formula(out, object, iterations, re_formula, verbose, ...)
@@ -426,6 +448,8 @@ pp_check.lm <- function(
   attr(out, "bandwidth") <- bandwidth
   attr(out, "model_info") <- minfo
   attr(out, "type") <- type
+  attr(out, "x_limits") <- x_limits
+
   class(out) <- c("performance_pp_check", "see_performance_pp_check", class(out))
   out
 }
@@ -440,6 +464,7 @@ pp_check.glm <- function(
   type = "density",
   verbose = TRUE,
   model_info = NULL,
+  x_limits = NULL,
   ...
 ) {
   # we need the formula and the response values to check for matrix responses
@@ -458,6 +483,7 @@ pp_check.glm <- function(
       type,
       verbose,
       model_info,
+      x_limits,
       ...
     ))
   }
@@ -527,6 +553,8 @@ pp_check.glm <- function(
   attr(out, "bandwidth") <- bandwidth
   attr(out, "model_info") <- minfo
   attr(out, "type") <- type
+  attr(out, "x_limits") <- x_limits
+
   class(out) <- c(
     "performance_pp_check",
     "see_performance_pp_check",